refactor: move ItemInfoFormValue into interface directory
