+++ src/domain/item/interface/inputData.ts
@@ -1,0 +1,10 @@
+export type ItemInfoFormValue = {
+  name: string
+  price: string
+  weight: string
+  stockRakuten: string
+  stockMakeshop: string
+  jancode: string
+}
+
+export type BlockValue = {}
